filter_rewrite: Moved `Teleflask` to `.extras.flask`. Maybe I could look into something like https://snarky.ca/lazy-importing-in-python-3-7/ ?
