+++ teleflask/server/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from .core import Teleserver
-from .extras import Teleflask
+from .extras.flask import Teleflask
 
 __author__ = 'luckydonald'
-
